+++ crates/crabe_framework/src/data/geometry/penalty.rs
@@ -4,9 +4,9 @@ use serde::Serialize;
 // TODO : Document
 #[derive(Serialize, Clone, Debug)]
 pub struct Penalty {
-    pub width: f32,
-    pub depth: f32,
-    pub top_left_position: Point2<f32>,
+    pub width: f64,
+    pub depth: f64,
+    pub top_left_position: Point2<f64>,
 }
 
 // TODO : Implement some helper methods
